requirements.txt
asgiref==3.3.1
braintree==4.5.0
certifi==2020.11.8
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.5.0
djangorestframework==3.12.2
idna==2.10
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
